split image post mixins into components
I have realised the styles I am doing are acting on classes that all
loops will have, and therefore are resuable for different types - a
split card with an image preview for a blog post, for instance.

I have moved the mixins to their own `components` module, with an index.
In the process I also refactored the mixins to make better use of
css-grid.